requirements.txt
torch
torchvision
transformers
sentence-transformers
faiss-cpu
umap-learn
plotly
streamlit
gradio
diffusers
